Avoid any action when the user runs scripts on an un-initialized CEPSparker folder

+++ Mac/clearPlayerDebugMode.command
@@ -9,27 +9,35 @@ export projectHomeDir=`dirname "$scriptDir"`
 
 cd "$projectHomeDir"
 
-export CEP_VERSION=`head -n 1 "$projectHomeDir/BuildSettings/CEPVersion.txt"`
+if [ -f  "$projectHomeDir/BuildSettings/CEPVersion.txt" ]; then
 
-if [ "$CEP_VERSION" == "4.2" ]; then
-  export PLIST_FILE=com.adobe.CSXS.4.plist
-elif [ "$CEP_VERSION" == "5.2" ]; then
-  export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ "$CEP_VERSION" == "6.1" ]; then
-  export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ "$CEP_VERSION" == "7.0" ]; then
-  export PLIST_FILE=com.adobe.CSXS.5.plist
-elif [ "$CEP_VERSION" == "8.0" ]; then
-  export PLIST_FILE=com.adobe.CSXS.8.plist
-fi
+    export CEP_VERSION=`head -n 1 "$projectHomeDir/BuildSettings/CEPVersion.txt"`
 
-if [ "$PLIST_FILE" != "" ]; then
+    if [ "$CEP_VERSION" != "" ]; then
+    
+        if [ "$CEP_VERSION" == "4.2" ]; then
+            export PLIST_FILE=com.adobe.CSXS.4.plist
+        elif [ "$CEP_VERSION" == "5.2" ]; then
+            export PLIST_FILE=com.adobe.CSXS.5.plist
+        elif [ "$CEP_VERSION" == "6.1" ]; then
+            export PLIST_FILE=com.adobe.CSXS.5.plist
+        elif [ "$CEP_VERSION" == "7.0" ]; then
+            export PLIST_FILE=com.adobe.CSXS.5.plist
+        elif [ "$CEP_VERSION" == "8.0" ]; then
+            export PLIST_FILE=com.adobe.CSXS.8.plist
+        else
+            echo "Unexpected CEP_VERSION"
+        fi
 
-  defaults write $PLIST_FILE PlayerDebugMode 0
-  defaults write $PLIST_FILE LogLevel 1
+        if [ "$PLIST_FILE" != "" ]; then
 
-  #
-  # Force Mac OS to re-read the PLIST file
-  #
-  killall -u `whoami` cfprefsd
+            defaults write $PLIST_FILE PlayerDebugMode 0
+            defaults write $PLIST_FILE LogLevel 1
+
+            #
+            # Force Mac OS to re-read the PLIST file
+            #
+            killall -u `whoami` cfprefsd
+        fi
+    fi
 fi
