+++ Mac/localDebugInstall.command
@@ -4,25 +4,33 @@
 
 export scriptDir=`dirname "$0"`
 cd $scriptDir
+
 export scriptDir=`pwd`
 export projectHomeDir=`dirname "$scriptDir"`
 
 cd "$projectHomeDir"
 
-export EXTENSION_DIRNAME=`head -n 1 "$projectHomeDir/BuildSettings/ExtensionDirName.txt"`
-export EXTENSION_HOMEDIR=~/Library/Application\ Support/Adobe/CEP/extensions/"$EXTENSION_DIRNAME"
+if [ -f "$projectHomeDir/BuildSettings/ExtensionDirName.txt" ]; then
 
-"$scriptDir/setPlayerDebugMode.command"
-"$scriptDir/adjustVersionInManifest.command"
+	export EXTENSION_DIRNAME=`head -n 1 "$projectHomeDir/BuildSettings/ExtensionDirName.txt"`
 
-rm -rf "$EXTENSION_HOMEDIR"
-mkdir "$EXTENSION_HOMEDIR"
+	if [ "$EXTENSION_DIRNAME" != "" ]; then
 
-ln -s "$projectHomeDir/debug" "$EXTENSION_HOMEDIR/.debug"
-ln -s "$projectHomeDir/css" "$EXTENSION_HOMEDIR/css"
-ln -s "$projectHomeDir/CSXS" "$EXTENSION_HOMEDIR/CSXS"
-ln -s "$projectHomeDir/html" "$EXTENSION_HOMEDIR/html"
-ln -s "$projectHomeDir/js" "$EXTENSION_HOMEDIR/js"
-ln -s "$projectHomeDir/jsx" "$EXTENSION_HOMEDIR/jsx"
-ln -s "$projectHomeDir/shared_js_jsx" "$EXTENSION_HOMEDIR/shared_js_jsx"
+		export EXTENSION_HOMEDIR=~/Library/Application\ Support/Adobe/CEP/extensions/"$EXTENSION_DIRNAME"
 
+		"$scriptDir/setPlayerDebugMode.command"
+		"$scriptDir/adjustVersionInManifest.command"
+
+		rm -rf "$EXTENSION_HOMEDIR"
+		mkdir "$EXTENSION_HOMEDIR"
+
+		ln -s "$projectHomeDir/debug" "$EXTENSION_HOMEDIR/.debug"
+		ln -s "$projectHomeDir/css" "$EXTENSION_HOMEDIR/css"
+		ln -s "$projectHomeDir/CSXS" "$EXTENSION_HOMEDIR/CSXS"
+		ln -s "$projectHomeDir/html" "$EXTENSION_HOMEDIR/html"
+		ln -s "$projectHomeDir/js" "$EXTENSION_HOMEDIR/js"
+		ln -s "$projectHomeDir/jsx" "$EXTENSION_HOMEDIR/jsx"
+		ln -s "$projectHomeDir/shared_js_jsx" "$EXTENSION_HOMEDIR/shared_js_jsx"
+		
+	fi
+fi
